Moved writers to separate module
