+++ program_structure/Cargo.toml
@@ -26,6 +26,7 @@ serde_derive = "1.0"
 serde-sarif = "0.3"
 serde_json = "1.0"
 thiserror = "1.0"
+termcolor = "1.1.3"
 
 [dev-dependencies]
 proptest = "1.0"
