+++ program_structure/src/utils/mod.rs
@@ -1,5 +1,5 @@
 pub mod constants;
 pub mod environment;
 pub mod nonempty_vec;
-pub mod report_writer;
+pub mod writers;
 pub mod sarif_conversion;
